+++ projects/core/src/routing/store/index.ts
@@ -1,5 +1,5 @@
 export * from './actions/index';
 export * from './effects/index';
 export * from './reducers/index';
-export * from './roting-state';
+export * from './routing-state';
 export * from './selectors/index';
