+++ projects/core/src/routing/store/reducers/router.reducer.spec.ts
@@ -9,7 +9,7 @@ import {
 } from '@ngrx/router-store';
 import { Store, StoreModule } from '@ngrx/store';
 import { PageType } from '../../../model/cms.model';
-import { RouterState } from '../roting-state';
+import { RouterState } from '../routing-state';
 import * as fromReducer from './router.reducer';
 
 @Component({
